projects/notion-api-test/requirements.txt
celery
notion-client
python-dotenv
celery[redis]